Create Coursera
This project focuses on creating an efficient solution in R for caching the inverse of a matrix, a computationally expensive operation. Matrix inversion is a fundamental task in linear algebra with wide applications in data analysis, statistical modeling, and machine learning. However, calculating the inverse repeatedly for the same matrix can lead to performance inefficiencies, particularly with large datasets or within iterative processes such as simulations or optimization algorithms.